Better spritesheet exporting
You can now choose how many columns you want your spritesheet to be, allowing you to export atlases.

ISSUE: If the number of columns is larger than the number of rows AND rows * columns are not equal to the number of frames, the result is weird. The exported image has transparent space on its bottom side.