Changed hint_tooltips from the old string formats to actual sentences
Besides the timeline buttons and color switch button being ready to have their changed shortcuts (when we implement them for these buttons) in their hint tooltips, there is no new functionality. This is merely to make the lives of translators easier.

This is the final commit of weird string IDs being replaced to actual sentences in Translations.pot