Append Global.root_directory to XDG_DATA_DIRS
This lets Pixelorama to search for brushes, palettes & patterns in the "pixelorama" folder that may be on the root folder of the executable. Useful for non-package builds, like itch.io, and for the using Pixelorama inside the editor. It should not cause issues for packages but if it does, let me know.