Fixes critical bug with saving projects that contained linked cels
It used to save the linked cel array which contained the canvas nodes into the .pxo. This, however, does not make sense, because the canvas nodes are being recreated when a .pxo is loaded. Therefore, they are different nodes than they used to be.

Now, an array containing the frame numbers which have cels linked for each frame is being saved and loaded into the pxo.