Small optimization in CanvasPreview
On large canvases (like 4096x4096), simply moving the mouse around was laggy. It turns out that texture_to_draw = current_cels[i].image_texture was causing the lag, but I'm not sure why. It could be that texture_to_draw was a Texture and not an ImageTexture.

There is still some room for optimization, both the second canvas and canvas preview are being drawn on _input(), which is unnecessary.