Rename store_info.md to extension_repository.md, add Voxelorama and remove other store links
The reason for removing the other store link is because I am worried that adding more links by default, can make it easier to compromise security. Extensions can be dangerous, and even if one link gets injected by malware, users are in danger, so I think it's better to keep the extensions is one, centralized place. I included all of the extensions listed in Variable's store link in this file.

I also renamed it from store to repository, as I think it's a more fitting name.