Remove the Manage Layouts dialog
Instead, add "Add Layout" and "Delete Layout" options in the layouts submenu. This makes layout management easier, more intuitive and quicker, as it now requires less clicks. This is similar to what Photoshop does with its Workspaces

By removing the manage layouts dialog, we also got rid of duplicate code that already existed in TopMenuContainer, and it should be less bug-prone now.