Make minimum cel size smaller
And allow it to become even smaller if the font size decreases. To avoid issues with the size of frame buttons being bigger than the size of the cel buttons on small sizes, their text is now being clipped, and on frames with three or more digits, their text is aligned on the right, to ensure that if they are clipped, the digits on the left are the ones who get clipped. For example, the text of the button of frame 131 would be clipped to 31.