Implement a basic hexagonal grid, remove the "All" option
This still needs improvements. Its lines are being drawn outside of the canvas, it needs optimizations and it needs to support the flat-top orientation.

The "All" option was used to draw both a rectangular and an isometric grid at the same time, but this is no longer needed now that we have multiple grids. I've also removed the isometric grid size field. Now there is one grid size for all types of grids.