Change all instances of extension "store" to extension "explorer" or "repository"
The word store may imply the existence of paid extensions. We only allow free & open source extensions on the explorer, both for security reasons and because hosting paid extensions would probably be a legal headache, so the term "Extension Explorer" is a better one. And instead of "store links", call them "repository links".