Hide transparent checker of palette swatches if their color is opaque
I did a stress test of instantiating 4096 swatches and Pixelorama was laggy. Turns out it was because of the too many TransparentCheckers being visible, making the GPU quite busy. Making them invisible totally solved the issue.

I don't know how much performance this actually saves in real world scenarios, but it might be a small performance boost for weaker GPUs.