In BaseDraw, move undo_redo.create_action to commit_undo
The reason it got placed to prepare_undo was to fix a bug where the user could undo in the middle of drawing, in #603. However, almost one year later, I introduced another fix for this issue which affected all tools, not just the design ones, in af2b1feb1f63144ebce00520ea2f8ee832dc49bd. Therefore the fix in #603 is no longer needed. I did some tests and nothing seems broken, but we should keep it in mind just in case.