Cancel tool on mobile devices if performing a touch screen gesture
Before, if a user pinched to zoom on mobile, the current left tool got activated for a few frames, until the system detected that the touch event is a gesture, and the image got drawn. Now, if the system detects a gesture, the tool get cancelled and the image reverts to its previous state.

In theory we could have just undo'd, but that would fill the undo history with unnecessary actions.