Add a button that lets users unlock the palette grid and remove PaletteScroll
Addresses #1356 by adding a button that unlocks the palette grid, allowing it to resize freely based on the panel size, instead of having a fixed width and height.

In order for this to work properly though, I had to remove the custom PaletteScroll script. We're now just adding all palette swatches as nodes in the grid and use Godot's ScrollContainer to handle the scroll logic, resulting in smoother scrolling, touch screen scrolling and cleaner code. The reason it was added was to save performance by avoiding instantiating so many nodes for each swatch, but in real life scenarios it isn't that big of a deal, and most palettes are small anyway. I also did some stress tests by instantiating thousands of swatch nodes and I didn't notice any impact in performance, besides increased memory usage. It only caused performance issues when the swatches also had their TransparentChecker node visible, but I disabled it for opaque colors for v1.1.6.