Download lospec palettes by pasting their URI
I wanted to quickly download a palette but I was too lazy to download the actual file. So I thought, what if I just copied the URI from the "Open In App" button and paste it in Pixelorama, since we already have a "Paste from clipboard" feature?

And so, this little feature was born. Out of my own laziness. You are welcome.