Add a single tool mode preference that makes the right tool the same as the left one
If this is enabled, the right mouse button will always activate the same tool as the left mouse button. This option is enabled by default on touch screen devices.

This mode has been on my mind for years now, and I finally decided to go ahead and implement it, now that we're so close to an Android version.