Implement gif importing and update gdgifexporter
Your prayers have been heard. Gif importing without needing FFMPEG is finally supported in Pixelorama. https://github.com/jegor377/godot-gdgifexporter/pull/25

In case a gif fails to load, FFMPEG is still used as a backup method, but I haven't been able to recreate such a case thus far.

It is not the fastest thing in the world, but it works for now. Eventually we may want to seriously consider C++ for both importing and exporting.