Remove unneeded line from Project.gd
This doesn't do anything, since the perspective editor has no _draw() method. From what I understand, it was supposed to call the "update" method, which is now "_update_points", and it got renamed to "queue_redraw" on #900. "_update_points" is now being called when the Global.project_switched signal gets emitted anyway, so it's not needed here.